feat: extração dos abstracts e organização da estrutura de pastas
