+++ langchain/main.py
@@ -1,0 +1,3 @@
+from data_extration.infos_extration import obter_abstract,obter_descricoes
+
+# Lógica de relacionar os dados retornados das listas
